% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macev_wew_doc.R
\docType{data}
\name{macev_wew_toelichting}
\alias{macev_wew_toelichting}
\title{Toelichting op de betekenis van de indicatoren van macev_wew_data}
\format{
Dataframe
\itemize{
\item \code{indicator_groep} Het type milieufactor
\item \code{indicator_waarde_code}  Code voor de milieufactor
\item \code{indicator_waarde} Waarde van de milieufactor
\item \code{omschrijving} Betekenis van de indicatorwaarde
}
}
\usage{
macev_wew_toelichting
}
\description{
Deze dataset bevat een toelichting op de betekenis van de indicatorwaarden in \link{macev_wew_data}.
}
\examples{

head(macev_wew_toelichting)
}
\seealso{
\link{macev_wew_data} en \link{macev_wew_zeldzaamheid}
}
\keyword{datasets}
