% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wew_mafa_doc.R
\docType{data}
\name{wew_mafa_data}
\alias{wew_mafa_data}
\title{Milieu- en habitatpreferenties van Nederlandse zoetwatermacrofauna}
\format{Dataframe met 4 kolommen
\itemize{
\item \code{taxonnaam} Taxonnaam
\item \code{indicator_groep} Naam van de milieufactor
\item \code{indicator_waarde} Waarde van de milieufactor
\item \code{indicator_waarde_aandeel} Relatief aandeel van de waarde binnen de milieufactor op een schaal van 0 tot 10.
}}
\source{
\url{http://www.wew.nu/publicaties/wew_themanummer23_soortenlijst.xls}

\url{http://www.wew.nu/publicaties/wew_themanummer23.pdf}
}
\usage{
wew_mafa_data
}
\description{
Deze dataset bevat autecologische informatie van Nederlandse macrofaunasoorten en is opgesteld door
de subgroep Autecologie van de \href{www.wew.nu}{Werkgroep Ecologische Waterbeheer}.
}
\details{
Voor negen factoren zijn de milieu- en habitatpreferenties van macrofaunasoorten verzameld,
namelijk zoutgehalte, zuurgraad, droogval, stroomsnelheid, substraat, diepte, oppervlakte,
saprobie en trofie. Deze factoren zijn gekozen omdat ze sterk sturend zijn en omdat er voldoende
gegevens van beschikbaar zijn. Voor elke factor zijn vier tot tien klassen onderscheiden.

Oorspronkelijk doel van de dataset
Bewerkingen + datum
}
\section{Metadata dataset}{

\itemize{
\item \strong{Naam:} Milieu- en habitatpreferenties van Nederlandse zoetwatermacrofauna
\item \strong{Bronhouder:} Werkgroep Ecologisch Waterbeheer
\item \strong{Auteurs:}
\itemize{
\item \strong{Auteurs:} W.C.E.P. Verberk, P.F.M. Verdonschot, T. van Haaren, B. van Maanen.
\item \strong{Redactie:} Rebi C. Nijboer.
\item \emph{Subgroep Autoecologie, WEW-23.}
}
\item \strong{Datum dataset:} Mei 2012.
\item \strong{Gebruiksrechten:} Geen beperkingen
\item \strong{Soortgroep:} Macrofauna
}
}

\seealso{
\link{wew_mafa_zeldzaamheid} en \link{wew_mafa_toelichting}
}
\keyword{datasets}
