% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wew_mafa_doc.R
\docType{data}
\name{wew_mafa_toelichting}
\alias{wew_mafa_toelichting}
\title{Toelichting op de betekenis van de indicatoren van wew_mafa_data}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 56 rows and 4 columns.}
\usage{
wew_mafa_toelichting
}
\description{
Deze dataset bevat een toelichting op de betekenis van de indicatorwaarden in \link{wew_mafa_data}.
}
\seealso{
\link{wew_mafa_data} en \link{wew_mafa_zeldzaamheid}
}
\keyword{datasets}
