% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wew_mafa_doc.R
\docType{data}
\name{wew_mafa_zeldzaamheid}
\alias{wew_mafa_zeldzaamheid}
\title{Zeldzaamheid van macrofauna}
\format{Dataframe met 4 kolommen
\itemize{
\item \code{taxonnaam} Taxonnaam
\item \code{zeldzaamheid} Code voor de zeldzaamheid
\item \code{omschrijving} Beschrijving van de zeldzaamheid
}}
\usage{
wew_mafa_zeldzaamheid
}
\description{
Deze dataset bevat de zeldzaamheid van Nederlandse macrofaunasoorten en is opgesteld door
de subgroep Autecologie van de \href{www.wew.nu}{Werkgroep Ecologische Waterbeheer}.
}
\seealso{
\link{wew_mafa_data} en \link{wew_mafa_toelichting}
}
\keyword{datasets}
